Adjust docs menu
