+++ docs/src/components/layout.js
@@ -6,8 +6,9 @@ import {
   Main,
   Box,
   Container,
+  useColorMode,
 } from 'theme-ui'
-import { useState } from 'react'
+import { useState, useRef } from 'react'
 import { Global } from '@emotion/core'
 
 import SkipLink from './skip-link'
@@ -16,9 +17,24 @@ import Footer from './footer'
 import Sidebar from './sidebar'
 import Pagination from './pagination'
 import MenuButton from './menu-button'
+import NavLink from './nav-link'
+import Button from './button'
 
-export default ({ header, ...props }) => {
+const modes = [
+  'light',
+  'dark',
+]
+
+export default props => {
   const [ menuOpen, setMenuOpen ] = useState(false)
+  const [ mode, setMode ] = useColorMode()
+  const nav = useRef(null)
+
+  const cycleMode = e => {
+    const i = modes.indexOf(mode)
+    const next = modes[(i + 1) % modes.length]
+    setMode(next)
+  }
 
   return (
     <Styled.root>
@@ -37,12 +53,24 @@ export default ({ header, ...props }) => {
       </SkipLink>
       <Layout>
         <Header>
-          {header}
           <MenuButton
             onClick={e => {
               setMenuOpen(!menuOpen)
+              if (!nav.current) return
+              const navLink = nav.current.querySelector('a')
+              if (navLink) navLink.focus()
             }}
           />
+          <NavLink to='/'>Theme UI</NavLink>
+          <Box mx='auto' />
+          <NavLink href='https://github.com/system-ui/theme-ui'>GitHub</NavLink>
+          <Button
+            css={{
+              ml: 2,
+            }}
+            onClick={cycleMode}>
+            {mode}
+          </Button>
         </Header>
         <Main>
           <Container
@@ -51,6 +79,7 @@ export default ({ header, ...props }) => {
               display: 'flex',
             }}>
             <Sidebar
+              ref={nav}
               open={menuOpen}
               onFocus={e => {
                 setMenuOpen(true)
